+++ js/core/forms.js
@@ -5,17 +5,28 @@
   var accounts = window.SalonAccountForms;
   var prestations = window.SalonPrestationForms;
   var misc = window.SalonMiscForms;
+  var absences = window.SalonAbsenceForms;
+  var holidays = window.SalonHolidayForms;
 
   window.SalonForms = {
+    absenceCard: absences.absenceCard,
+    bindAbsenceCardActions: absences.bindAbsenceCardActions,
+    bindHolidayCardActions: holidays.bindHolidayCardActions,
     bindPhotoInput: misc.bindPhotoInput,
     bindPrestationActions: prestations.bindPrestationActions,
     bindReservationCardActions: reservations.bindReservationCardActions,
     cancelReservation: reservations.cancelReservation,
+    canManageAbsence: absences.canManageAbsence,
     configure: formState.configure,
+    deleteAbsence: absences.deleteAbsence,
     deleteAccount: accounts.deleteAccount,
-    openAbsenceForm: misc.openAbsenceForm,
+    deleteHoliday: holidays.deleteHoliday,
+    holidayCard: holidays.holidayCard,
+    holidaysForCollab: holidays.holidaysForCollab,
+    openAbsenceForm: absences.openAbsenceForm,
     openAddAccountForm: accounts.openAddAccountForm,
     openClientForm: clients.openClientForm,
+    openHolidayForm: holidays.openHolidayForm,
     openPrestationForm: prestations.openPrestationForm,
     openProfileForm: accounts.openProfileForm,
     openReservation: reservations.openReservation,
@@ -23,6 +34,9 @@
     reservationCard: reservations.reservationCard,
     resetLink: accounts.resetLink,
     resetPassword: accounts.resetPassword,
-    setReservationStatus: reservations.setReservationStatus
+    saveAbsence: absences.saveAbsence,
+    saveHoliday: holidays.saveHoliday,
+    setReservationStatus: reservations.setReservationStatus,
+    toggleAccountActive: accounts.toggleAccountActive
   };
 }());